separated header element from Main Page
