+++ selenium_tests/page_objects/MainPage.py
@@ -5,18 +5,6 @@ from .BasePage import BasePage
 
 class MainPage(BasePage):
     NAVIGATION_BAR_ITEMS = (By.XPATH, "//ul[contains(@class, 'navbar-nav')]/li/a")
-    MY_ACCOUNT_DROPDOWN = (By.CSS_SELECTOR, "a[title='My Account']")
-    REGISTER_USER_LINK_REL = (By.XPATH, "//following-sibling::ul//a[text()='Register']")
-    LOGIN_AS_USER_LINK_REL = (By.XPATH, "//following-sibling::ul//a[text()='Login']")
-    SEARCH_FIELD = (By.NAME, "search")
-    SEARCH_BUTTON = (By.CSS_SELECTOR, "#search button")
-    CART_BUTTON = (By.CSS_SELECTOR, ".btn-inverse")
-    YOUR_STORE_LINK = (By.CSS_SELECTOR, "[id='logo'] > a")
-    CURRENCY_DROPDOWN = (By.XPATH, "//button/span[text()= 'Currency']/ancestor::button")
-    CURRENCY_ICON = (By.XPATH, "//button/span[text()= 'Currency']/preceding-sibling::strong")
-    EUR_CURRENCY_ITEM = (By.XPATH, "//form[@id='form-currency']//button[@name='EUR']")
-    GBP_CURRENCY_ITEM = (By.XPATH, "//form[@id='form-currency']//button[@name='GBP']")
-    USD_CURRENCY_ITEM = (By.XPATH, "//form[@id='form-currency']//button[@name='USD']")
 
     MENU_ITEMS_NAMES = ['Desktops', 'Laptops & Notebooks', 'Components', 'Tablets', 'Software',
                         'Phones & PDAs', 'Cameras', 'MP3 Players']
@@ -56,42 +44,3 @@ class MainPage(BasePage):
 
     def click_nav_menu_item(self, menu_item: str):
         self._click(self.get_menu_item_element_by_name(menu_item))
-
-    def expand_my_account_dropdown(self):
-        self._click(self.MY_ACCOUNT_DROPDOWN)
-
-    def click_register_account(self):
-        self._click_child_element(self.MY_ACCOUNT_DROPDOWN, self.REGISTER_USER_LINK_REL)
-
-    def click_login_into_user_account(self):
-        self._click_child_element(self.MY_ACCOUNT_DROPDOWN, self.LOGIN_AS_USER_LINK_REL)
-
-    def input_text_into_search_field(self, search_text: str):
-        self._input_field_value(self.SEARCH_FIELD, search_text)
-
-    def click_search_button(self):
-        self._click(self.SEARCH_BUTTON)
-
-    def click_cart_button(self):
-        self._click(self.CART_BUTTON)
-
-    def click_your_store_link(self):
-        self._click(self.YOUR_STORE_LINK)
-
-    def expand_currency_dropdown(self):
-        self._click(self.CURRENCY_DROPDOWN)
-        return self
-
-    def select_currency(self, currency_value: str):
-        """Select currency from currency drop-down and verify value selected"""
-        if currency_value == "EUR":
-            self._select_dropdown_item(self.EUR_CURRENCY_ITEM)
-            self._verify_text_in_element(self.CURRENCY_ICON, "€")
-        elif currency_value == "GBP":
-            self._select_dropdown_item(self.GBP_CURRENCY_ITEM)
-            self._verify_text_in_element(self.CURRENCY_ICON, "£")
-        elif currency_value == "USD":
-            self._select_dropdown_item(self.USD_CURRENCY_ITEM)
-            self._verify_text_in_element(self.CURRENCY_ICON, "$")
-        else:
-            raise AssertionError(f"{currency_value} currency is not supported")
